requirements.txt
# AI Policy Co-Pilot MVP Requirements - UPDATED FOR WEAVIATE
# Core FastAPI and web framework
fastapi>=0.100.0
uvicorn>=0.23.0

# Database drivers
neo4j>=5.0.0
sqlalchemy>=2.0.0

# Vector database and embeddings (UPDATED FOR WEAVIATE)
weaviate-client==4.5.1  # NEW - Replaces psycopg2-binary and pgvector
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0  # Optional - for local fallback
chromadb>=0.4.0   # Optional - for experimentation

# Data processing and NLP
pandas>=2.0.0
numpy>=1.24.0
spacy>=3.7.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0

# PDF processing and OCR
PyMuPDF>=1.23.0
pdfplumber>=0.9.0
pytesseract>=0.3.10
easyocr>=1.7.0
pypdf2>=3.0.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
playwright>=1.40.0

# LangChain and RAG
langchain>=0.1.0
langchain-community>=0.0.20
llama-index>=0.9.0

# Legal NLP models
# Note: These may need to be installed separately
# InLegalBERT: https://huggingface.co/law-ai/InLegalBERT
# RelBERT: https://huggingface.co/RelBERT/legal

# Streamlit UI
streamlit>=1.28.0
streamlit-option-menu>=0.3.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pydantic>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: Cloud and deployment
docker>=6.1.0
docker-compose>=1.29.0

# Optional: LLM APIs
openai>=1.0.0
anthropic>=0.3.0